+++ src/DotnetDiagnosticsMcp.Server/Hosting/DiagnosticServiceRegistration.cs
@@ -104,6 +104,7 @@ internal static class DiagnosticServiceRegistration
         services.AddSingleton<DotnetDiagnosticsMcp.Core.Memory.IInvestigationSummaryExporter, DotnetDiagnosticsMcp.Core.Memory.InvestigationSummaryExporter>();
         services.AddSingleton<DotnetDiagnosticsMcp.Core.Memory.ISummaryComparer, DotnetDiagnosticsMcp.Core.Memory.SummaryComparer>();
         services.AddSingleton<DotnetDiagnosticsMcp.Core.Memory.IMemoryTrendCollector, DotnetDiagnosticsMcp.Core.Memory.MemoryTrendCollector>();
+        services.AddSingleton<DotnetDiagnosticsMcp.Core.ProcessDiscovery.IRuntimeConfigInspector, DotnetDiagnosticsMcp.Core.ProcessDiscovery.RuntimeConfigInspector>();
         services.AddSingleton<DotnetDiagnosticsMcp.Core.ProcessDiscovery.IProcessResourcesCollector, DotnetDiagnosticsMcp.Core.ProcessDiscovery.ProcessResourcesCollector>();
         services.AddSingleton<DotnetDiagnosticsMcp.Core.ProcessDiscovery.IRequestsNowCollector, DotnetDiagnosticsMcp.Core.ProcessDiscovery.RequestsNowCollector>();
         services.AddSingleton<DotnetDiagnosticsMcp.Core.Drilldown.IDiagnosticHandleStore>(_ =>